requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
Werkzeug>=2.3.0
python-dateutil>=2.8.0
